Edited parameters start location to be at 0 0 0 Changed ConfigurationManager to set all floats instead of ints Added the configurationFile initialization in the Main Added the use of the config file in Robot CPP

+++ MyRobot/ConfigurationManager.cpp
@@ -71,37 +71,37 @@ string ConfigurationManager::getMap()
 	return this->map;
 }
 
-int ConfigurationManager::getStartLocationX()
+float ConfigurationManager::getStartLocationX()
 {
 	return this->startLocationX;
 }
 
-int ConfigurationManager::getStartLocationY()
+float ConfigurationManager::getStartLocationY()
 {
 	return this->startLocationY;
 }
 
-int ConfigurationManager::getStartLocationYaw()
+float ConfigurationManager::getStartLocationYaw()
 {
 	return this->startLocationYaw;
 }
 
-int ConfigurationManager::getGoalX()
+float ConfigurationManager::getGoalX()
 {
 	return this->goalX;
 }
 
-int ConfigurationManager::getGoalY()
+float ConfigurationManager::getGoalY()
 {
 	return this->goalY;
 }
 
-int ConfigurationManager::getRobotSizeX()
+float ConfigurationManager::getRobotSizeX()
 {
 	return this->robotSizeX;
 }
 
-int ConfigurationManager::getRobotSizeY()
+float ConfigurationManager::getRobotSizeY()
 {
 	return this->robotSizeY;
 }
